Config overhaul, build script
