+++ .gitignore
@@ -1,2 +1,3 @@
 **/bin
 **/bck
+**/build
