+++ include.txt
@@ -1,0 +1,1 @@
+*.p3d;*.paa
